update: writeFileSync

+++ utils/template.js
@@ -1,19 +1,25 @@
 const handlebars = require('handlebars')
 const Metalsmith = require('metalsmith')
 const fs = require('fs')
+const path = require('path')
 
 const Tpl = {
   initPackageJson: (projectInfo) => {
-    const packageObj = {
-      name: projectInfo.name,
-      description: projectInfo.description,
-      author: projectInfo.author
+    
+    try {
+      console.log('to initPackageJson')
+      const fileName = path.join(projectInfo.name, 'package.json')
+      console.log(fileName)
+      if (fs.existsSync(fileName)) {
+        const content = fs.readFileSync(fileName).toString()
+        const result = handlebars.compile(content)(projectInfo)
+        console.log('result', result)
+        // 重写覆盖
+        fs.writeFileSync(fileName, result)
+      }
+    } catch (err) {
+      console.log('err', err)
     }
-    const fileName = `${packageObj.name}/package.json`
-    const content = fs.readFileSync(fileName).toString()
-    const result = handlebars.compile(content)(packageObj)
-    // 重写覆盖
-    fs.writeFileSync(fileName, result)
   },
 
   /**
@@ -32,11 +38,12 @@ const Tpl = {
         Object.keys(files).forEach(fileName => {
           const fileText = files[fileName].contents.toString()
           if (fileName === 'package.json') {
+            console.log('------')
             Tpl.initPackageJson(projectInfo)
           } else {
             files[fileName].contents = Buffer.from(handlebars.compile(fileText)(meta));
             // fs.writeFileSync(fileName, files[fileName].contents)
-            console.log(files[fileName].contents.toString())
+            // console.log(files[fileName].contents.toString())
           }
         })
         // package_temp_content = handlebars.compile(fs.readFileSync(packjsonTemp).toString())(meta);
